feat: merge tests and move the constant to index
